Fixed nasty bug renaming FuzzyFlags

+++ app/src/main/java/com/stack/gocode/localData/fuzzy/Defuzzifier.java
@@ -8,27 +8,27 @@ import com.stack.gocode.localData.Named;
 
 public class Defuzzifier implements Named {
     private String name;
-    private double speed1, speed2;
+    private int speed1, speed2;
 
-    public Defuzzifier(String name, double speed0, double speed1) {
+    public Defuzzifier(String name, int speed0, int speed1) {
         this.name = name;
         this.speed1 = speed0;
         this.speed2 = speed1;
     }
 
     public String getName() {return name;}
-    public double getSpeed1() {return speed1;}
-    public double getSpeed2() {return speed2;}
+    public int getSpeed1() {return speed1;}
+    public int getSpeed2() {return speed2;}
 
     public void setName(String newName) {
         this.name = newName;
     }
 
-    public void setSpeed1(double newSpeed) {
+    public void setSpeed1(int newSpeed) {
         this.speed1 = newSpeed;
     }
 
-    public void setSpeed2(double newSpeed) {
+    public void setSpeed2(int newSpeed) {
         this.speed2 = newSpeed;
     }
